Reducing cyclomatic complexity further, extracting helper class to improve readability
